Renamed algorithms.h -> algorithm.h
