+++ inst/include/Rcpp.h
@@ -81,5 +81,5 @@
 #include <Rcpp/platform/solaris.h>
 #include <Rcpp/api/meat/meat.h>
 
-#include <Rcpp/algorithms.h>
+#include <Rcpp/algorithm.h>
 #endif
