+++ inst/include/Rcpp/algorithm.h
